requirements.txt
black==24.8.0
click==8.1.7
coverage==7.6.1
customtkinter==5.2.2
darkdetect==0.8.0
exceptiongroup==1.2.2
flake8==7.1.1
ghp-import==2.1.0
gitchangelog==3.0.4
importlib_metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
-e git+https://github.com/rsekijima/pdf_gen.git@a9bc8cfd78dc5092066dae93f4ee4c13932501e8#egg=pdf_gen
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyyaml_env_tag==0.1
six==1.16.0
tk==0.1.0
tomli==2.0.2
typing_extensions==4.12.2
watchdog==4.0.2
zipp==3.20.2
